lesson5_django/my_first_project/requirements.txt
django==1.11.5
mysqlclient==1.3.12
gunicorn==19.7.1
djangorestframework==3.6.4
requests
djangorestframework-jwt